refactor: custom-dsl 분리
