+++ src/main/kotlin/com/petqua/domain/product/review/ProductReviewCustomRepositoryImpl.kt
@@ -1,9 +1,6 @@
 package com.petqua.domain.product.review
 
-import com.linecorp.kotlinjdsl.dsl.jpql.Jpql
 import com.linecorp.kotlinjdsl.dsl.jpql.jpql
-import com.linecorp.kotlinjdsl.dsl.jpql.sort.SortNullsStep
-import com.linecorp.kotlinjdsl.querymodel.jpql.predicate.Predicate
 import com.linecorp.kotlinjdsl.render.jpql.JpqlRenderContext
 import com.linecorp.kotlinjdsl.render.jpql.JpqlRenderer
 import com.petqua.common.util.createQuery
@@ -26,7 +23,7 @@ class ProductReviewCustomRepositoryImpl(
         paging: ProductReviewPaging
     ): List<ProductReviewWithMemberResponse> {
 
-        val query = jpql {
+        val query = jpql(ProductReviewDynamicJpqlGenerator) {
             selectDistinctNew<ProductReviewWithMemberResponse>(
                 entity(ProductReview::class),
                 entity(Member::class),
@@ -49,24 +46,5 @@ class ProductReviewCustomRepositoryImpl(
             jpqlRenderer,
             paging.limit
         )
-    }
-
-    fun Jpql.productReviewIdLt(lastViewedId: Long?): Predicate? {
-        return lastViewedId?.let { path(ProductReview::id).lt(it) }
-    }
-
-    fun Jpql.photoOnlyEq(photoOnly: Boolean): Predicate? {
-        return if (photoOnly) path(ProductReview::hasPhotos).eq(true) else null
-    }
-
-    fun Jpql.productReviewScoreEq(score: Int?): Predicate? {
-        return score?.let { path(ProductReview::score).eq(it) }
-    }
-
-    fun Jpql.sortBy(sorter: ProductReviewSorter): SortNullsStep? {
-        return when (sorter) {
-            ProductReviewSorter.REVIEW_DATE_DESC -> path(ProductReview::createdAt).desc()
-            ProductReviewSorter.RECOMMEND_DESC -> path(ProductReview::recommendCount).desc()
-        }
     }
 }
